fix batching for scalar values
